backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
numpy==1.24.3
pandas==2.0.3
scikit-image==0.21.0
joblib==1.3.2
ultralytics==8.0.196
Pillow==10.0.0